improve performance and readability for ZeBadgeViewModel (#171)
