+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zevm/MapExtensions.kt
@@ -1,0 +1,12 @@
+package de.berlindroid.zeapp.zevm
+
+fun <K, V> Map<K, V>.copy(vararg entries: Pair<K, V>): Map<K, V> {
+    val result = toMutableMap()
+
+    entries.forEach { entry ->
+        val (replaceKey: K, replaceValue: V) = entry
+        result[replaceKey] = replaceValue
+    }
+
+    return result.toMap()
+}
