+++ zeapp/app/src/main/java/de/berlindroid/zeapp/zevm/ZeBadgeViewModel.kt
@@ -30,9 +30,9 @@ import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
+import timber.log.Timber
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
-import timber.log.Timber
 import javax.inject.Inject
 
 private const val MESSAGE_DISPLAY_DURATION = 3_000L
@@ -94,7 +94,7 @@ class ZeBadgeViewModel @Inject constructor(
 
         messageProgressJob?.cancel()
         messageProgressJob = viewModelScope.launch {
-            (0 until 10).forEach { progress ->
+            for (progress in 0 until 10) {
                 _uiState.update {
                     it.copy(messageProgress = 1.0f - progress / MESSAGE_DISPLAY_UPDATES.toFloat())
                 }
@@ -206,40 +206,31 @@ class ZeBadgeViewModel @Inject constructor(
             }
         } else {
             // no selection needed, check for name slot and ignore non configurable slots
-            val newCurrentSlotEditor: ZeEditor?
             val slots = _uiState.value.slots
-            when (slot) {
-                is ZeSlot.Name -> {
-                newCurrentSlotEditor = ZeEditor(
+            val newCurrentSlotEditor = when (slot) {
+                is ZeSlot.Name -> ZeEditor(
                     slot,
                     slots[ZeSlot.Name]!!,
                 )
-                }
 
-                is ZeSlot.QRCode -> {
-                newCurrentSlotEditor = ZeEditor(
+                is ZeSlot.QRCode -> ZeEditor(
                     slot,
                     slots[ZeSlot.QRCode]!!,
                 )
-                }
 
-                is ZeSlot.Weather -> {
-                newCurrentSlotEditor = ZeEditor(
+                is ZeSlot.Weather -> ZeEditor(
                     slot,
                     slots[ZeSlot.Weather]!!,
                 )
-                }
 
-                is ZeSlot.BarCode -> {
-                newCurrentSlotEditor = ZeEditor(
+                is ZeSlot.BarCode -> ZeEditor(
                     slot,
                     slots[ZeSlot.BarCode]!!,
                 )
-                }
 
                 else -> {
-                newCurrentSlotEditor = null
                     Timber.d("Customize Page", "Cannot configure slot '${slot.name}'.")
+                    null
                 }
             }
             newCurrentSlotEditor?.let { currentSlotEditor ->
@@ -455,17 +446,6 @@ class ZeBadgeViewModel @Inject constructor(
             currentSimulatorSlot = ZeSlot.Name,
             slots = emptyMap(),
         )
-}
-
-private fun <K, V> Map<K, V>.copy(vararg entries: Pair<K, V>): Map<K, V> {
-    val result = toMutableMap()
-
-    entries.forEach { entry ->
-        val (replaceKey: K, replaceValue: V) = entry
-        result[replaceKey] = replaceValue
-    }
-
-    return result.toMap()
 }
 
 data class ZeBadgeUiState(
